Move bot file to a new package
